Prvni pokus o log akce. Kratky stisk cudlu noviny je strucne info, dlouhy stisk je s detailama. Prihlaseni do webu prestehovano na dlouhy stisk sklenice do dalsich akci, protoze casem by to melo nacitat carovy kod z obrazovky.

+++ app/src/main/java/com/scurab/barcodescanner2/forest/ItemdView.java
@@ -18,7 +18,7 @@ public class ItemdView implements Serializable {
     // objem
     public double Volume;
     // cena
-    public int ConsPrice;
+    public double ConsPrice;
     // konzument
     public String Username;
     // Zkonzumováno kdy
